Edit for practice MR - contoroller, model. main.go - change file names
