+++ domain/model/sushi.go
@@ -1,3 +1,10 @@
 package model
 
 // ここにmodelの情報を定義する
+
+type Sushi struct {
+	ID    int    `json:"id"`
+	Name  string `json:"name"`
+	Price int    `json:"price"`
+	Stock int    `json:"stock"`
+}
