+++ main.go
@@ -4,18 +4,15 @@ import (
 	"log"
 	"net/http"
 
-	controller "github.com/Sushiro-gacha/sushiro-gacha-api/app/controller"
+	"github.com/Sushiro-gacha/sushiro-gacha-api/app/controller"
 )
 
-// func HandllerHello() {
-// 	return practicecontroller.RequestHello()
-// }
 func main() {
 	StartMainServer()
 }
 
 func StartMainServer() {
-	http.HandleFunc("/gacha/price/", controller.GathaPrice)
+	http.HandleFunc("/gacha/price/", controller.GachaPrice)
 	//http.HandleFunc("/home/calorie/", controller.GachaCalories)
 
 	log.Fatal(http.ListenAndServe(":8000", nil))
